Beacon now fires in a different way, rather than having an argument list it just passes an options dictionary, it's how things are done these days and it a better replacement that what I was using.
Also Beacon now provides the same "hittest" that Kapture does, that way I can use Beacon ans my event fallthrough rather than Kapture, this makes it so that views capture events rather than key combos, which is better in almost every way. (for the user and the developer).

The application object now has a "fire_event" (slightly different than the "fire" of Beacon) that will take care of firing the event at the right object, that way you can fire all the events you want and the application object will take care of the rest.

Moved most of the navigation logic from the main.js and added into the GAListView controller, which is much cleaner and should interfere with others less.